requirements.txt
pyTelegramBotAPI==4.14.0
aiohttp>=3.8.0
openai==1.3.7
requests==2.31.0
python-dateutil==2.8.2
python-docx==1.1.1
python-dotenv
Pillow>=10.1.0
reportlab==4.0.4
httpx==0.25.2
pdf2image==1.16.3
PyPDF2==3.0.1
